Don't abort on missing component.md or .scss files
Previously, missing component.md (the ones with the extra info, like the
components' descriptions) or component.scss files aborted the generation
of components docs. But as the remaining info (mainly from the Vue
files) already give much information, we don't want to abort the whole
process but rather log a warning (in case of a missing component.md) and
continue anyway, accepting that the generated component docs files
aren't 100% complete.
This allows for better incremental completion of the components docs.

Furthermore, a missing component.scss file is no error case but there
are components without special styling. In this case, no warning is
logged.