Add GlideJS to SCSS import resolution
Simply adding it to the includePath isn't enough here as the compiler
has problems with webpack alias "~". Hence, before passing the line with
the import statement to the SASS compiler, we remove everything that
isn't needed, leaving only the directory part which isn't already in the
includePath.
That way, both webpack and the direct SASS compiler call are fine.