Update test for SfHero
Multiple places in the structure for the component changed, that's why
the test needed an overhaul. We make use of mock items now, so the slots
get correct data which was previously passed via props. Also the binding
of the "go" method changed, which also required a more complicated test
setup via scoped slots.