Insert storybook basic story into components docs
This inserts an iframe into the final components docs. It shows the
content of the storybook iframe of the "basic" story of a component. As
the content is served from a different domain, it is not possible to
determine the perfect size of the frame. Hence, this size has to be
determined manually and added to the components descriptive MD file (new
headline: "storybook-iframe").

At this state, no code example is rendered yet. This functionality will
be added in a later commit.