Generate common usage section in component docs from stories files
Instead of writing all common usage sections for our components docs
manually, this enables auto-generation of the example code blocks by
using the "common" stories in the storybook.

As parsing via regex is no viable solution--everything can happen inside
a story! it might have external dependencies referenced, or contain
characters that look like the end of a story although they're only part
of a string. So, the story has to be actually executed. Luckily, the
story is defined by a "story function" ("storyFn") which describes the
template, props and components references. All that is passed to a
"storiesOf" function, more strictly the ".add()" function of its return
type. As this is JS land, we can simply create a closure, so the
executed story will use our functions instead of the defined imports
inside the story file--those are simply removed before passing the story
file content to eval().

Extra work had to be done for the props: Those aren't simple objects we
can access from the storyFn but function calls. Those can be proxied via
previously closure functions too, so those are collected and later on
concatenated to build the final code block.

For the code block to look nicely formatted, prettier's API is used.

Side note: I disabled ESLint's no-unused-vars rule for the part where
the evaluated story code magic happens because it looks to the linter as
if it's unused. Using WebStorm you will still see many warnings in that
area because WS doesn't provide a way to disable the internal linter for
a specific block or section (the only other way is to use comments for
each and every line with "unused" definitions are to disable the
inspection completely. Both ways are dumb.)