Add stories decorators
This new file contains decorators which can be used for Storybook
stories.
Some decorators accept parameters (e.g. withMargin) and can be called
like normal functions:
  .addDecorator(withMargin("1rem"))

Others have default values and can be added without an actual call:
  .addDecorator(withCenteredContainer)

A special case is the "withDataToggle" decorator: It adds a toggle
(SfCheckbox) which binds to a data property which propagates from the
outer decorator component through the Storybook component to the actual
component which holds the data. Likewise events from the actual
component "bubble up" to the outer decorator component and show state
changes in realtime.
For this to work, the component also needs a mixin which reflect those
state changes in the component.

This approach is useful for future issues like #480. It doesn't pollute
the stories file. All we have to do there to add e.g. visibility toggles
is to add the decorator and the mixin.

For convenience reasons the generic dataToggle-decorator and -mixin come
also in a specific version for the "visible" data key.