Make components docs aware of stories decorators
This makes the components docs aware of the data-mixins in our stories:
Although those stories do not include the actual "data" property (e.g.
for "visible" or "isSidebarOpen") because those are dynamically added
by the decorator+mixin logic, we need this information for the users who
read our components docs (without the data field the template wouldn't
make sense), hence this logic introduced here adds those fields to the
components docs by analyzing the dataToggleMixin.
(More mixins can be added later to the auto-generation scripts, when/if
we introduce more mixins like this.)