feat(organisms): add SfMegaMenu (#472)
* Cleaned stories and styles

* linter

* Created template for mega-menu content

* Linter error

* Finished styling for desktop

* Litn fix

* Linter fix

* lint

* linter

* cleaning properties

* Working on custom class modifiers

* Set multiple, removed arrows

* Removed transition

* Styling

* Removed accordion

* Megamenu - working on refactor

* Added missing method

* Implemented pages for megamenu

* New template for megamenu

* Cleaned up scss

* Styling changes

* Styling changes for mobile, fixing tests

* Styling fix for dekstop

* Insert storybook basic story into components docs

This inserts an iframe into the final components docs. It shows the
content of the storybook iframe of the "basic" story of a component. As
the content is served from a different domain, it is not possible to
determine the perfect size of the frame. Hence, this size has to be
determined manually and added to the components descriptive MD file (new
headline: "storybook-iframe").

At this state, no code example is rendered yet. This functionality will
be added in a later commit.

* Remove storybookTable generation

We don't provide component docs inside Storybook anymore; those moved to
the official docs.

* Remove default exports and dummy prop knob from stories

* Use Storybook's Group-ID feature to group props and css modifiers

And run prettier.
(Ignore whitespace changes to see relevant changes)

* Rename first storybook story to "Common"

This story will be used in the auto-generation of the common usage
section in the components docs.

* Fix typo in SfBannerGrid

* Use real prop names in storybook knobs (no hyphens)

* Add new line in component docs template for successful VuePress render

* Remove custom common usage declaration from existing component MDs

Those will be generated from the storybook. But the component MDs will
define values for the specific iframe heights. These values here were
tested manually to fit the content.

* Remove unused storybook-notes.md

* Generate common usage section in component docs from stories files

Instead of writing all common usage sections for our components docs
manually, this enables auto-generation of the example code blocks by
using the "common" stories in the storybook.

As parsing via regex is no viable solution--everything can happen inside
a story! it might have external dependencies referenced, or contain
characters that look like the end of a story although they're only part
of a string. So, the story has to be actually executed. Luckily, the
story is defined by a "story function" ("storyFn") which describes the
template, props and components references. All that is passed to a
"storiesOf" function, more strictly the ".add()" function of its return
type. As this is JS land, we can simply create a closure, so the
executed story will use our functions instead of the defined imports
inside the story file--those are simply removed before passing the story
file content to eval().

Extra work had to be done for the props: Those aren't simple objects we
can access from the storyFn but function calls. Those can be proxied via
previously closure functions too, so those are collected and later on
concatenated to build the final code block.

For the code block to look nicely formatted, prettier's API is used.

Side note: I disabled ESLint's no-unused-vars rule for the part where
the evaluated story code magic happens because it looks to the linter as
if it's unused. Using WebStorm you will still see many warnings in that
area because WS doesn't provide a way to disable the internal linter for
a specific block or section (the only other way is to use comments for
each and every line with "unused" definitions are to disable the
inspection completely. Both ways are dumb.)

* styling fixes

* Fix for images

* Update packages/vue/src/components/organisms/SfMegaMenu/SfMegaMenu.stories.js

Co-Authored-By: Przemysław Spaczek <pspaczek@divante.co>

* Update packages/vue/src/components/organisms/SfMegaMenu/SfMegaMenu.html

Co-Authored-By: Przemysław Spaczek <pspaczek@divante.co>

* Update packages/vue/src/components/organisms/SfMegaMenu/SfMegaMenu.html

Co-Authored-By: Przemysław Spaczek <pspaczek@divante.co>

* Update packages/vue/src/components/organisms/SfMegaMenu/SfMegaMenu.js

Co-Authored-By: Przemysław Spaczek <pspaczek@divante.co>

* Removed unused prop

* Update packages/vue/src/components/organisms/SfMegaMenu/SfMegaMenu.html

Co-Authored-By: Przemysław Spaczek <pspaczek@divante.co>

* Update packages/vue/src/components/organisms/SfMegaMenu/SfMegaMenu.html

Co-Authored-By: Przemysław Spaczek <pspaczek@divante.co>

* Renamed slots

* Update packages/vue/src/components/organisms/SfMegaMenu/_internal/SfMegaMenuColumn.html

Co-Authored-By: Przemysław Spaczek <pspaczek@divante.co>

* Fixed issue with mobile

* Added logic for mobile

* Replace Node12 syntax with compatible syntax for older versions

* Added vertical padding

* Added keys

* Use Netlify preview for iframe links in this PR

This will be reverted once this PR is approved and before it is merged.

* Fix for render issue

* Renamed to updateItems

* Image fix

* Added docs

* small refactor SfFooter

* small refactor SfMegaMenu

* Migrate new components SfFooter and SfHeader to new stories and docs format

* Fix prop names in SfCircleIcon stories

* Add missing group IDs to story knobs

* Correct comment

* fix test

* fix style

* add background

* add SfBar for SfMegaMenu

* remove v-model

* fix test

* Lint after merging

* Created file for custom decorator

* Added centered decorator for SfSection

* add layout

* layout cd.

* fix

* clean SfMegaMenu stories

* clean watchers handlers

* fix SfImage for picture

* fix style for empty aside slot

* clean SfMegaMenuStories

* fix uppercase menu

* fix mega menu behaviour for desktop

* update es-lint

* integration POC

* remove SfHeaderNavigationItem

* fix SfFooter stories

* typo

* update SfHeader docs

* fix SfHeader logo style

* update MegaMenu docs

* lint

* remove Layout example

* remove unnecessary SfFooter method

* fix SfFooterColumn slot description

* use provide/inject

* rename and split parentMenu to visible and title props

* update docs

* fix unit test

* back to _internal SfHEaderNavigationItem

* fix watcher

* move SfMegaMenu/SfHeader integration from Layout to SfMegaMenu stories

* fix: component stories

* feat: add SfMegaMenuItem

* chore: prepare to merge

* fix: remove unused SfIcon import

* refactor: small changes

* refactor: position

* refactor: classes

* fix: bottom border on nav items

* fix: docs

Co-authored-by: Przemek Spaczek <pspaczek@divante.co>
Co-authored-by: Eduard Dopler <kontakt@eduard-dopler.de>